Fix resuming.

+++ exporters/export_managers/base_exporter.py
@@ -84,8 +84,9 @@ class BaseExporter(object):
         for bypass_script in self.bypass_cases:
             try:
                 bypass_script.meets_conditions()
-                self.persistence.close()
-                self.persistence.delete()
+                if not self.config.exporter_options.get('resume'):
+                    self.persistence.close()
+                    self.persistence.delete()
                 self.bypass_exporter(bypass_script)
                 return True
             except RequisitesNotMet:
